backend/requirements.txt
fastapi
uvicorn
websockets>=13.0
python-dotenv
google-genai
numpy
opencv-python-headless
pydantic
